requirements.txt
langchain[openai]
langchain-community
opencv-python
pydantic
PyYAML
pillow
numpy
